Force reload the view on start
This is basically a giant hammer to fix issues with the custom element
sometimes not showing up, probably a race in Lovelace.

If the custom element isn't defined when the parent card is generated, a
<hui-error-card> is inserted in its place (from `create-element-base.ts`)
and a `ll-rebuild` event is fired if the custom element is registered
before a timeout.  <hui-view> is expected to listen for that event (on the
card holding the footer that contains the custom element) and rebuild the
card; however, this seems to fail sometimes and nothing is rendered
instead.  Likely the footer element was created but not inserted into the
DOM (because it was created on `setConfig()`, but only inserted into the
DOM after `render()` is processed).